fixes after Merle's user test make category icons dependent on user (before in contacts tab system was trying to match own category icons to foreign categories) expand all gets marginleft auto because justify content does not work on mobile devices add a description what the "comment" field in meals is for change photo upload (MIME "image" does not work on mobile, now try "image/*"

+++ src/components/Meals/AddMeal.jsx
@@ -68,7 +68,9 @@ const AddMeal = (props) => {
   return (
     <>
       <Navbar pageTitle={t('New Meal')} leftSideComponent={<BackButton onClick={() => {
-        deleteAllImagesFromMeal(meal._id);
+        if (meal.images) {
+          deleteAllImagesFromMeal(meal._id);
+        }
         history.goBack();
       }} />} rightSideComponent={meal.title ? <DoneButton onClick={addNewMeal} /> : null} />
